+++ addon/query/predicate.js
@@ -80,6 +80,72 @@ export class SimplePredicate extends BasePredicate {
 }
 
 /**
+ * The class of date predicate for filtering attribute by value and filter operator.
+ *
+ * @namespace Query
+ * @class DatePredicate
+ * @extends BasePredicate
+ *
+ * @param attributePath {String} The path to the attribute for filtering.
+ * @param operator {Query.FilterOperator|String} The filter operator.
+ * @param value {String|Date} The value for filtering.
+ * @constructor
+ */
+export class DatePredicate extends BasePredicate {
+  constructor(attributePath, operator, value) {
+    super();
+
+    this._attributePath = attributePath;
+    this._operator = FilterOperator.tryCreate(operator);
+    this._value = value;
+  }
+
+  /**
+   * The path to the attribute for filtering.
+   *
+   * @property attributePath
+   * @type String
+   * @public
+   */
+  get attributePath() {
+    return this._attributePath;
+  }
+
+  /**
+   * The filter operator.
+   *
+   * @property operator
+   * @type Query.FilterOperator
+   * @public
+   */
+  get operator() {
+    return this._operator;
+  }
+
+  /**
+   * The value for filtering.
+   *
+   * @property value
+   * @type String
+   * @public
+   */
+  get value() {
+    return this._value;
+  }
+
+  /**
+   * Converts this instance to string.
+   *
+   * @method toString
+   * @return {String} Text representation of the predicate.
+   * @public
+   */
+  toString() {
+    return `(${this._attributePath} ${this._operator} ${this._value})`;
+  }
+}
+
+/**
  * The class of complex predicate which include multiple predicates unioned with logical condition.
  *
  * @namespace Query
@@ -202,6 +268,64 @@ export class StringPredicate extends BasePredicate {
    */
   contains(value) {
     this._containsValue = value;
+    return this;
+  }
+}
+
+/**
+ * The predicate class for geography attributes.
+ *
+ * @namespace Query
+ * @class GeographyPredicate
+ * @extends BasePredicate
+ *
+ * @param {String} attributePath The path to the attribute for predicate.
+ * @constructor
+ */
+export class GeographyPredicate extends BasePredicate {
+  constructor(attributePath) {
+    super();
+
+    if (!attributePath) {
+      throw new Error('Attribute path is required for GeographyPredicate constructor.');
+    }
+
+    this._attributePath = attributePath;
+    this._intersectsValue = null;
+  }
+
+  /**
+   * The path to the attribute for predicate.
+   *
+   * @property attributePath
+   * @type {String}
+   * @public
+   */
+  get attributePath() {
+    return this._attributePath;
+  }
+
+  /**
+   * The geography value that has to intersect with the attribute.
+   *
+   * @property intersectsValue
+   * @type {String}
+   * @public
+   */
+  get intersectsValue() {
+    return this._intersectsValue;
+  }
+
+  /**
+   * Sets the value that the attribute has to contain.
+   *
+   * @method contains
+   * @param {String} geography The geography value that has to intersect with the attribute.
+   * @return {Query.StringPredicate} Returns this instance.
+   * @chainable
+   */
+  intersects(geography) {
+    this._intersectsValue = geography;
     return this;
   }
 }